finish test_loading
